saku-strategy/backend/requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
asgiref==3.9.1
django-cors-headers==4.3.1
blis==0.7.11
catalogue==2.0.10
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.8
cloudpathlib==0.22.0
confection==0.1.5
cryptography==45.0.7
cymem==2.0.11
dateparser==1.2.0
Django==4.2.24
djangorestframework==3.15.2
exceptiongroup==1.3.0
idna==3.10
immutabledict==4.2.1
iniconfig==2.1.0
Jinja2==3.1.6
langcodes==3.5.0
language_data==1.3.0
lxml==6.0.1
marisa-trie==1.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.13
numpy==1.26.4
ortools==9.10.4067
packaging==25.0
pandas==2.2.2
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.3.0
pluggy==1.6.0
preshed==3.0.10
protobuf==6.32.0
psycopg2-binary==2.9.9
PuLP==2.8.0
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pypdfium2==4.30.0
pytest==8.3.2
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
rich==14.1.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
thinc==8.2.5
tomli==2.2.1
tqdm==4.67.1
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.3
PyJWT==2.8.0
djangorestframework-simplejwt==5.3.0
